+++ fedora20-py3-x64.sh
@@ -5,9 +5,12 @@ set -e
 sudo yum install git nano openssl-devel
 sudo yum groupinstall "Development Tools"
 
-curl -O https://www.python.org/ftp/python/3.3.3/Python-3.3.3.tgz
+rm -Rf Python-3.3.3
+
+if [[ ! -e Python-3.3.3.tgz ]]; then
+    curl -O https://www.python.org/ftp/python/3.3.3/Python-3.3.3.tgz
+fi
 tar xvfz Python-3.3.3.tgz
-rm Python-3.3.3.tgz
 
 cd Python-3.3.3
 ./configure
