+++ ubuntu1004-py3-x64.sh
@@ -4,9 +4,12 @@ set -e
 
 sudo apt-get install git-core build-essential libssl-dev
 
-curl -O https://www.python.org/ftp/python/3.3.3/Python-3.3.3.tgz
+rm -Rf Python-3.3.3
+
+if [[ ! -e Python-3.3.3.tgz ]]; then
+    curl -O https://www.python.org/ftp/python/3.3.3/Python-3.3.3.tgz
+fi
 tar xvfz Python-3.3.3.tgz
-rm Python-3.3.3.tgz
 
 cd Python-3.3.3/
 ./configure
